refactor: изменено расположение некоторых классов
