+++ src/main/java/com/warehouse/myshop/product/service/ProductService.java
@@ -1,7 +1,7 @@
 package com.warehouse.myshop.product.service;
 
 
-import com.warehouse.myshop.product.dto.FilterConditionDto;
+import com.warehouse.myshop.product.dto.condition.FilterConditionDto;
 import com.warehouse.myshop.product.dto.ListProductDto;
 import com.warehouse.myshop.product.dto.NewProductDto;
 import com.warehouse.myshop.product.dto.ResponseProductDto;
